Warn on trivial_numeric_casts

+++ quinn-udp/src/unix.rs
@@ -544,7 +544,7 @@ fn recv(io: SockRef<'_>, bufs: &mut [IoSliceMut<'_>], meta: &mut [RecvMeta]) -> 
         }
     };
     for i in 0..(msg_count as usize) {
-        meta[i] = decode_recv(&names[i], &hdrs[i], hdrs[i].msg_datalen as usize)?;
+        meta[i] = decode_recv(&names[i], &hdrs[i], hdrs[i].msg_datalen)?;
     }
     Ok(msg_count as usize)
 }
@@ -621,6 +621,7 @@ fn prepare_msg(
     if is_ipv4 {
         if !sendmsg_einval {
             #[cfg(not(target_os = "netbsd"))]
+            #[expect(trivial_numeric_casts)]
             {
                 encoder.push(libc::IPPROTO_IP, libc::IP_TOS, ecn as IpTosTy);
             }
@@ -769,7 +770,7 @@ fn decode_recv(
             (libc::IPPROTO_IPV6, libc::IPV6_PKTINFO) => {
                 let pktinfo = unsafe { cmsg::decode::<libc::in6_pktinfo, libc::cmsghdr>(cmsg) };
                 dst_ip = Some(IpAddr::V6(Ipv6Addr::from(pktinfo.ipi6_addr.s6_addr)));
-                interface_index = Some(pktinfo.ipi6_ifindex as u32);
+                interface_index = Some(pktinfo.ipi6_ifindex);
             }
             #[cfg(any(target_os = "linux", target_os = "android"))]
             (libc::SOL_UDP, gro::UDP_GRO) => unsafe {
